% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amVennDiagram.R
\name{amVennDiagram}
\alias{amVennDiagram}
\title{Venn diagram widget}
\usage{
amVennDiagram(
  data,
  theme = "default",
  legendPosition = "right",
  elementId = NULL
)
}
\arguments{
\item{data}{a list such as one returned by \code{\link{makeVennData}}}

\item{theme}{the theme: \code{"default"}, \code{"dark"}, \code{"dataviz"},
\code{"frozen"}, \code{"kelly"}, \code{"material"}, \code{"moonrise"},
or \code{"spirited"}}

\item{legendPosition}{legend position: \code{"right"} or \code{"bottom"}}

\item{elementId}{a HTML id (usually useless)}
}
\value{
An \code{amVennDiagram} widget.
}
\description{
Creates an \code{amVennDiagram} widget.
}
\examples{
sets <- list(A = 1:20, B = 10:30, C = 15:35)
dat <- makeVennData(sets)
amVennDiagram(dat, theme = "kelly")
}
