% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeVennData.R
\name{makeVennData}
\alias{makeVennData}
\title{Venn diagram data from a list of sets}
\usage{
makeVennData(sets)
}
\arguments{
\item{sets}{a named list of vectors representing some sets}
}
\value{
A list suitable for usage in \code{\link{amVennDiagram}}.
}
\description{
Make data for usage in \code{\link{amVennDiagram}}.
}
\examples{
sets <- list(A = 1:20, B = 10:30, C = 15:35)
dat <- makeVennData(sets)
amVennDiagram(dat, theme = "spirited")
}
